Initialized LSS texture recognizer example
Initial example recognizes horizontal and vertical stripes on
monochromatic image using right and bottom neighbours of pixel.

Here is illustration:
- current pixel X
- right neighbour R
- bottom neighbour B
 __ __
|X |R |
|__|__|
|B |
|__|

They can be 0 either 1.

Amount of textures is 2: vertical | and horizontal -
||--
||--
||--
So, to represent pixel as a point, it's enough to have a vector of
length: (neighbourhood types) * (colors amount)^2 + 1, where
- neighborhood types are right and bottom neighbourhood, which is 2
- colors amount is 2
We need an ability to store all neighbours of each pixel with any colors
in one vector to apply Linear Separator, and 1 more component for
constant, which is 1 for any vector.

If we have a pair (colorX, colorY) for neighbourhood N, corresponding
component of current vector should be 1, and so on for each neighbour of
pixel X.
Last element is 1 by previous agreement.
All other values are 0.